Differentiate the shape of Worker and Fighter for better visibility
Also factored out agent_class.rs
